+++ test/test_conv.c
@@ -2,10 +2,10 @@
 #include "../include/datahandler/datahandler.h"
 
 int main(){
-    double kernel_weight[27] = {-1, 0, 1, -2, 0, 2, -1, 0, 1,
+    float kernel_weight[27] = {-1, 0, 1, -2, 0, 2, -1, 0, 1,
                                 -1, 0, 1, -2, 0, 2, -1, 0, 1,
                                 -1, 0, 1, -2, 0, 2, -1, 0, 1};
-    double kernel_bias[] = {0};
+    float kernel_bias[] = {0};
     struct Kernel kernel;
     kernel.row = 3;
     kernel.col = 3;
